+++ docs/changelog.md
@@ -12,6 +12,9 @@ Requires libvips v8.7.4.
 * Add `composite` operation supporting multiple images and blend modes; deprecate `overlayWith`.
   [#728](https://github.com/lovell/sharp/issues/728)
 
+* Add support for `pages` input option for multi-page input.
+  [#1566](https://github.com/lovell/sharp/issues/1566)
+
 * Add support for `page` input option to GIF and PDF.
   [#1595](https://github.com/lovell/sharp/pull/1595)
   [@ramiel](https://github.com/ramiel)